% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgainsim.R
\name{p_gain_quantile_fit}
\alias{p_gain_quantile_fit}
\title{Computation of log-linear fit of the pgain-quantiles (dependent on the number of tests) and evaluation for a determined number of tests}
\usage{
p_gain_quantile_fit(
  pgain_types = c("add", "rec"),
  pgain_quantile,
  n_data_ff = nrow(pgain_quantile[[1]]),
  start_vec = lapply(1:length(pgain_types), function(x) data.frame(matrix(rep(c(1,
    0.01, 1.01), times = ncol(pgain_quantile[[1]])), ncol = ncol(pgain_quantile[[1]]),
    dimnames = list(row.names = c("a", "b", "d"), col.names =
    colnames(pgain_quantile[[1]]))))),
  test_number
)
}
\arguments{
\item{pgain_types}{Character vector of the types of p-gains, of which the quantiles were computed in the function p_gain_quantiles. ("add" means additive p-gain, "rec" means recessive p-gain, "dom" means dominant p-gain) (default = c("add","rec")).}

\item{pgain_quantile}{list of data frames. For every type of p-gain there is a data frame (in the same order as they are listed in pgain_types). Columns describe pgain-quantiles for different allele frequenciey (numeric values) and rows discribe number of tests. Output of function p_gain_quantiles.}

\item{n_data_ff}{Integer. Number of quantile datapoints that should be used for the plot. n_data_ff is a divider of the number of available datapoints (default = nrow(pgain_quantile[[1]])).}

\item{start_vec}{list of data frames. For every type of p-gain there is a data frame (in the same order as they are listed in pgain_types). Columns describe 3 starting estimates (numeric values) for the log-linear fit for each allele frequency (default is c(1,0.01,1.01) for every p-gain type and for every allele frequency).}

\item{test_number}{Integer. Number of tests for which the p-gain threshold should be determined.}
}
\value{
list_fits as a list, plots of log-linear fit of the quantiles for every allele frequency, approximated quantile for test_number many tests for every allele frequency. The list contains for every p-gain type a list with the log-linear fits for every allele frequency.
}
\description{
Computation of log-linear fit of the pgain-quantiles (dependent on the number of tests) and evaluation for a determined number of tests
}
\examples{
		
sim_data <- p_gain_simulation(pgain_types=c("add","rec"), AFs=c(0.1,0.5),n=10000L,snps_per_trait=1L,n_study=1000L,cores=2L)   
pgain_quantile <- p_gain_quantiles(pgain_types=c("add","rec"), n_tests=50L,sim_data)
list_fits <- p_gain_quantile_fit(pgain_types=c("add","rec"),pgain_quantile,n_data_ff=nrow(pgain_quantile[[1]]),start_vec=list(data.frame(c(1.09,0.012,1.014),c(1.12,0.014,1.019)),data.frame(c(1.09,0.012,1.014),c(1.12,0.014,1.019))),test_number=200L)

}
