% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgainsim.R
\name{p_gain_simulation}
\alias{p_gain_simulation}
\title{Random draw of recessive (and dominant) p-gains under no association.}
\usage{
p_gain_simulation(
  pgain_types = c("add", "rec"),
  AFs,
  n = 100000L,
  n_study = 1000L,
  cores = 2L
)
}
\arguments{
\item{pgain_types}{Character vector of the types of p-gains, which should be simulated. "add" means additive p-gain, "rec" means recessive p-gain, "dom" means dominant p-gain (default = c("add","rec")).}

\item{AFs}{Numeric vector of assumed allele frequencies.}

\item{n}{Integer. The number of random draws.}

\item{n_study}{Integer. The number of samples per simulation / study size (default = 1000L).}

\item{cores}{Integer. Amount of CPU cores used (<=1 : sequential)}
}
\value{
list_pgains_AF list of data frames. For every type of p-gain there is a data frame (in the same order as they are listed in pgain_types) . Each column describes simulated p-gain-values for the allele frequency, which is in the column name.
}
\description{
To simulate additive, recessive and dominant p-gains under no association traits are generated by random draws of a standard normally distributed sample of size n_study.
The genotype data is simulated in Hardy-Weinberg equilibrium (independently from trait data) as corrected Bernoulli distributed samples of size n_study with probability depending on the fixed allele frequencies AFs.
The p-values of the additive, recessive and dominat model are computed by applying linear regression with the additive, recessive and dominant dosage respectively as independent variable on the simulated trait and genotype data.
Based on the p-values of the three different models the additive, recessive and the dominant p-gains are computed for every allele frequency of AFs.
}
\examples{
sim_dat <- p_gain_simulation(pgain_types=c("add","rec"),AFs=c(0.3,0.7),n=10000L, cores=4L)

}
