% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgainsim.R
\name{p_gain_quantiles}
\alias{p_gain_quantiles}
\title{Computation of p-gain-quantiles for numbers of tests based on the result of function p_gain_simulation.}
\usage{
p_gain_quantiles(n_tests, sim_data)
}
\arguments{
\item{n_tests}{Integer. The number of tests for which the pgain-quantile should be computed. It depends on the available number of datapoints.}

\item{sim_data}{data frame. Each column describes simulated p-gain-values for the allele frequency, which is in the column name. Output of function p_gain_simulation.}
}
\value{
p_gain_quant data frame. Columns describe pgain-quantiles for different allele frequenciey (numeric values) and rows discribe number of tests.
}
\description{
Computation of p-gain-quantiles for numbers of tests based on the result of function p_gain_simulation.
}
\examples{
sim_data <- p_gain_simulation(AFs=c(0.1,0.5),n=10000L,snps_per_trait=1L,n_study=1000L,cores=2L)   # Beispiele müssen self contained sein
pgain_quantile <- p_gain_quantiles(n_tests=50L,sim_data)

}
